Added sneshacking library and started integrating into the code because it's honestly a better idea than refactoring the code and testing it when the code has already been testing so I'll likely be removing my version of it with due time

+++ src/Application/Data/Overworld.cc
@@ -10,7 +10,7 @@ using namespace Core;
 using namespace Graphics;
 
 Overworld::~Overworld() {
-  for (int i = 0; i < tiles32.size(); i++) {
+  for (int i = 0; i < (int) tiles32.size(); i++) {
     free(allmapsTilesLW[i]);
     free(allmapsTilesDW[i]);
     free(allmapsTilesSP[i]);
